BAH-2049 | refactor. moved tests to home component
