+++ src/constants.ts
@@ -15,7 +15,6 @@ export const attrName = 'practitioner_type'
 export const attrValue = 'Doctor'
 export const orderStatusCompleted = 'COMPLETED'
 export const loggedInUserKey = 'loggedInUser'
-export const isAuditLogPostedKey = 'isAuditLogPosted'
 export const isAuditLogEnabledKey = 'isAuditLogEnabled'
 export const successMessage = 'Report successfully uploaded'
 export const failureMessage = 'Failed to upload report'
