+++ src/home-button/home-button.tsx
@@ -1,51 +1,10 @@
-import React, {useEffect, useState} from 'react'
-import styles from './home-button.scss'
 import Home24 from '@carbon/icons-react/lib/home/24'
-import {
-  homePath,
-  isAuditLogEnabledKey,
-  isAuditLogPostedKey,
-  loggedInUserKey,
-} from '../constants'
 import {HeaderGlobalAction} from 'carbon-components-react'
-import {
-  auditLogGlobalPropertyURL,
-  auditLogURL,
-  fetcher,
-  getPayloadForUserLogin,
-  postApiCall,
-} from '../utils/api-utils'
-import {getCurrentUser, LoggedInUser} from '@openmrs/esm-framework'
-import useSWR from 'swr'
-
-interface AuditLogResponse {
-  data: boolean
-}
+import React from 'react'
+import {homePath} from '../constants'
+import styles from './home-button.scss'
 
 const HomeButton = () => {
-  const [loggedInUser, setLoggedInUser] = useState<LoggedInUser | null>(null)
-
-  let {data: auditLogEnabledResponse, error: auditLogResponseError} = useSWR<
-    AuditLogResponse,
-    Error
-  >(auditLogGlobalPropertyURL, fetcher)
-
-  useEffect(() => {
-    const subscription = getCurrentUser({
-      includeAuthStatus: false,
-    }).subscribe(setLoggedInUser)
-    return () => subscription.unsubscribe()
-  }, [])
-
-  if (auditLogEnabledResponse?.data && loggedInUser) {
-    const ac = new AbortController()
-    const auditMessagePayload = getPayloadForUserLogin(loggedInUser.username)
-    postApiCall(auditLogURL, auditMessagePayload, ac)
-    localStorage.setItem(loggedInUserKey, loggedInUser.username)
-    localStorage.setItem(isAuditLogPostedKey, 'true')
-    localStorage.setItem(isAuditLogEnabledKey, 'true')
-  }
-
   return (
     <HeaderGlobalAction
       aria-label="Home"
